% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCResultMethods.r
\name{MCResult.calcPaBaTiesRatio}
\alias{MCResult.calcPaBaTiesRatio}
\alias{calcPaBaTiesRatio}
\title{Calculate PaBa Ties Ratio.}
\usage{
MCResult.calcPaBaTiesRatio(.Object)
}
\arguments{
\item{.Object}{object of class "MCResult".}
}
\value{
PaBa ties ratio (ties slopes vs total slopes)
}
\description{
This function computes the ratio of slopes ties values in the classic pairwise
PaBa slope calculation. A ratio higher than approx 0.05 may suggest a potential
bias risk. In this case data precision is too low for correct PaBa estimation.
The function is written in R and can be slow with large data set.
}
