% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCResultResamplingMethods.r
\name{MCResultResampling.plotBoxEllipses}
\alias{MCResultResampling.plotBoxEllipses}
\title{Plot Box Ellipses of bootstrap coefficients}
\usage{
MCResultResampling.plotBoxEllipses(.Object, robust.cov = "MCD")
}
\arguments{
\item{.Object}{Object of class "MCResultResampling"}

\item{robust.cov}{Method for covariance. Default "MCD"}

\item{breaks}{see function 'hist' (?hist) for details}

\item{...}{further graphical parameters}
}
\value{
No return value
}
\description{
Plot Box Ellipses of bootstrap coefficients (slope and intercept).
}
