% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCResultBCaMethods.r
\name{MCResultBCa.plotBoxEllipses}
\alias{MCResultBCa.plotBoxEllipses}
\title{Plot Box Ellipses of bootstrap coefficients}
\usage{
MCResultBCa.plotBoxEllipses(.Object, robust.cov = "MCD")
}
\arguments{
\item{.Object}{Object of class "MCResultResampling"}

\item{robust.cov}{Method for covariance. Default "MCD"}
}
\value{
No return value
}
\description{
Plot Box Ellipses of bootstrap coefficients (slope and intercept).
}
